# fix: Fixes and optimizes NameVerification and ProfanityProtection