--- a/Projects/UOContent/Misc/NameVerification.cs
+++ b/Projects/UOContent/Misc/NameVerification.cs
@@ -6,16 +6,6 @@ namespace Server.Misc;
 
 public static class NameVerification
 {
-    public static readonly SearchValues<char> AlphaNumeric = SearchValues.Create(
-        '0', '1', '2', '3', '4', '5', '6', '7', '8', '9',
-        'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J',
-        'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T',
-        'U', 'V', 'W', 'X', 'Y', 'Z',
-        'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j',
-        'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't',
-        'u', 'v', 'w', 'x', 'y', 'z'
-    );
-
     public static readonly SearchValues<char> Alphabetic = SearchValues.Create(
         'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J',
         'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T',